feat(wasm): remove JSPI back-end entirely — Phase 7 of #2452
The pthread + SharedArrayBuffer coroutine back-end has been the default
since #2457 (Phase 6) and works in every cross-origin-isolated webview.
JSPI was kept around as an opt-out for one release; this commit removes
it so there is a single, simpler code path.

Source removed:
  - src/platforms/wasm/coroutine_platform_wasm.js.cpp.hpp  (EM_JS scheduler)
  - src/platforms/wasm/coroutine_platform_wasm.hpp (JSPI ICoroutinePlatform)

The pthread implementation moves to the canonical name:
  src/platforms/wasm/coroutine_platform_wasm_pthread.hpp →
  src/platforms/wasm/coroutine_platform_wasm.hpp

Conditional compilation gates removed:
  - All `#ifdef FASTLED_WASM_PTHREADS` in coroutine_runtime_wasm.impl.hpp
    are gone — pthread is the only path. The pthread suspendMainthread()
    / wakeWaiters() overrides are now unconditional.
  - `-DFASTLED_WASM_PTHREADS=1` is no longer emitted (no consumer left).

Build pipeline simplified:
  - ci/wasm_flags.py: drop `_jspi_explicitly_requested`,
    `_pthreads_enabled`, `_JSPI_FLAG_PREFIXES`,
    `_apply_pthread_substitution`, and the `os` import (~88 lines).
    The TOML loader now returns flags as-authored.
  - ci/wasm_build.py: drop `FASTLED_WASM_JSPI` / `FASTLED_WASM_PTHREADS`
    from both `_LINK_ENV_VARS` and the source fingerprint env loop —
    they no longer mutate the build.
  - build_flags.toml: drop `-sJSPI` + `-sJSPI_EXPORTS=…` from
    [linking.base]; fold the pthread link flags (`-pthread`,
    `-sPROXY_TO_PTHREAD`, `-sPTHREAD_POOL_SIZE=4`) directly into
    [linking.base]; drop the now-empty [linking.pthread] section.
    Add `-pthread` to [all].compiler_flags so every TU is built with it.

Verified end-to-end:
  - bash lint: clean
  - bash compile wasm --examples Blink: success
  - uv run python ci/wasm_compile.py examples/wasm: success
  - uv run ci/wasm_test.py: headless Chromium reports
    "controller running, worker active" — pthread back-end renders frames
  - bash test --cpp: 304/304 host unit tests pass
  - Smoke test of flag loader: no -sJSPI in link, no -DFASTLED_WASM_PTHREADS
    define emitted, -pthread + -sPROXY_TO_PTHREAD + -sPTHREAD_POOL_SIZE=4
    present in link flags

Closes #2452 (final phase).
