+++ src/platforms/coroutine.impl.cpp.hpp
@@ -12,7 +12,7 @@
 ///   - ICoroutineTask::exitCurrent()
 ///
 /// Supported platforms:
-/// - WASM: JSPI-based cooperative coroutines via coroutine_runtime_wasm.impl.hpp
+/// - WASM: pthread + SharedArrayBuffer cooperative coroutines via coroutine_runtime_wasm.impl.hpp
 /// - Stub/Host: std::thread implementation via coroutine_stub.impl.hpp (for testing)
 /// - ESP32: FreeRTOS tasks via coroutine_esp32.impl.hpp
 /// - Teensy: Cooperative ARM context switching via coroutine_teensy.impl.hpp
@@ -26,7 +26,7 @@
 #include "platforms/wasm/is_wasm.h"
 
 #if defined(FL_IS_WASM)
-// WASM: JSPI-based cooperative coroutines
+// WASM: pthread + SharedArrayBuffer cooperative coroutines
 #include "platforms/wasm/coroutine_runtime_wasm.impl.hpp"
 #elif defined(FASTLED_STUB_IMPL)
 // Stub/Host: Thread-based coroutine runner
