+++ src/platforms/wasm/compiler/build_flags.toml
@@ -56,21 +56,20 @@ compiler_flags = [
     # Language Standard and Compatibility
     "-std=c++20",                               # Use C++20 standard (enables header unit imports)
     "-fpermissive",                             # Allow some non-conforming code patterns
-    
+
     # Warning Configuration
     "-Wno-constant-logical-operand",            # Suppress constant logical operand warnings
     "-Wnon-c-typedef-for-linkage",             # Warn about non-C typedef issues
     "-Werror=bad-function-cast",               # Error on bad function casts
     "-Werror=cast-function-type",              # Error on function type cast issues
     "-Werror=unused-private-field",            # Unused private fields are dead code
-    
 
-    
     # Thread Safety and Runtime Behavior
+    "-pthread",                                 # Enable Emscripten pthreads (SAB + Web Workers) — required by the cooperative coroutine back-end (issue #2452)
     "-fno-threadsafe-statics",                 # Disable thread-safe static initialization (performance)
     "-fno-exceptions",                         # Disable C++ exceptions for all targets (size/performance)
     "-fno-rtti",                              # Disable RTTI for all targets (size/performance, consistent)
-    
+
     # C++20 Header Unit Support
     "-Wno-experimental-header-units",         # Suppress experimental header unit warning
 ]
@@ -309,45 +308,12 @@ flags = [
     # Linker Performance Optimization
     "-fuse-ld=wasm-ld",                      # Explicitly use wasm-ld (default, but explicit for clarity)
 
-    # Coroutine back-end. The DEFAULT is now the pthread + SharedArrayBuffer
-    # back-end (see [linking.pthread] below). ci/wasm_flags.py removes the
-    # -sJSPI* flags from this block at link time and substitutes the pthread
-    # block, plus adds -pthread + -DFASTLED_WASM_PTHREADS=1 to every compile
-    # unit. The legacy JSPI back-end is kept as an opt-out and is enabled by
-    # setting FASTLED_WASM_JSPI=1 in the environment. See issue #2452.
-    #
-    # JSPI (JavaScript Promise Integration) — enables cooperative coroutines
-    # via engine-level stack management. Unlike asyncify, JSPI requires NO code
-    # transformation and NO function whitelist — zero WASM size overhead.
-    # The engine creates/suspends/resumes stacks natively via WebAssembly.promising.
-    # Supported in Chrome 123+, Firefox 139+. NOT supported in WebKit (Safari,
-    # WKWebView, WebKitGTK) — that's the motivation for the pthread default.
-    #
-    # JSPI_EXPORTS: WASM exports that can transitively call a suspending import
-    # (EM_ASYNC_JS). Any export that reaches contextSwitch → _jspi_context_switch
-    # must be listed. EM_ASYNC_JS imports are handled automatically.
-    "-sJSPI",
-    "-sJSPI_EXPORTS=['main','extern_loop','extern_setup','jspi_run_entry']",
-]
-
-[linking.pthread]
-# ================================================================================================
-# PTHREAD-BACK-END LINKING FLAGS  (default since #2452 phase 6)
-# ================================================================================================
-# These flags are applied by ci/wasm_flags.py automatically when the pthread
-# back-end is selected (which is now the default). The loader strips the
-# -sJSPI / -sJSPI_EXPORTS flags from [linking.base] above and appends the
-# flags listed here instead. It also threads -pthread +
-# -DFASTLED_WASM_PTHREADS=1 into every compile unit so
-# fl::thread / fl::mutex / fl::condition_variable resolve to real backing
-# primitives (SharedArrayBuffer + Atomics.wait under the hood) and the
-# pthread variant of ICoroutinePlatform gets registered.
-#
-# To opt OUT of pthread and use the legacy JSPI back-end instead, set the
-# environment variable FASTLED_WASM_JSPI=1 before running `bash compile wasm`.
-# See issue #2452.
-
-flags = [
+    # Coroutine back-end — pthreads + SharedArrayBuffer. Cooperative semantics
+    # are provided by fl::thread + fl::mutex + fl::condition_variable in
+    # src/platforms/wasm/coroutine_platform_wasm.hpp. Works in every
+    # cross-origin-isolated webview (WebKit/Safari, WebKitGTK, WebView2,
+    # Firefox, Chromium). See issue #2452 for the migration history from
+    # JSPI to pthreads.
     "-pthread",                              # Enable Emscripten pthreads (SAB + Web Workers)
     "-sPROXY_TO_PTHREAD",                    # Run main() on a worker pthread so it can block
     "-sPTHREAD_POOL_SIZE=4",                 # Pre-allocate 4 workers (tune after benchmarking)
