+++ src/platforms/wasm/coroutine_platform_wasm.hpp
@@ -3,118 +3,169 @@
 // IWYU pragma: private
 
 /// @file coroutine_platform_wasm.hpp
-/// @brief WASM coroutine platform using JSPI (JavaScript Promise Integration)
+/// @brief WASM coroutine platform using pthreads + SharedArrayBuffer
 ///
-/// Implements ICoroutinePlatform using JSPI for cooperative context switching
-/// in WASM. Unlike asyncify, JSPI handles stack management at the engine level
-/// with zero WASM code transformation — no whitelist needed, no size overhead.
+/// ICoroutinePlatform implementation that uses OS-level threads (Emscripten
+/// pthreads on top of Web Workers + SharedArrayBuffer) plus a
+/// mutex/condition_variable baton to provide cooperative context switching.
 ///
-/// Each coroutine runs as a separate JSPI-promising WASM call. Context switching
-/// suspends the caller via an EM_ASYNC_JS import and resumes the target by
-/// resolving its pending Promise. A lightweight JS scheduler coordinates this.
+/// Each coroutine context owns a real OS thread that is parked on its own
+/// condition variable while another coroutine runs. contextSwitch hands a
+/// "baton" from the current thread to the target thread, then blocks the
+/// caller — only one thread runs user code at a time, so existing sketch
+/// code stays single-owner.
 ///
-/// Requires -sJSPI linker flag. Supported in Chrome 123+.
+/// Works in every cross-origin-isolated webview (Safari/WKWebView, Firefox,
+/// Chrome, WebView2, WebKitGTK). The previous JSPI back-end was removed in
+/// #2452 phase 7 — see that issue for the migration history.
 
 #include "platforms/wasm/is_wasm.h"
 
-// JSPI platform is compiled out when the pthread back-end is selected.
-// See platforms/wasm/coroutine_platform_wasm_pthread.hpp.
-#if defined(FL_IS_WASM) && !defined(FASTLED_WASM_PTHREADS)
+#ifdef FL_IS_WASM
 
 // IWYU pragma: begin_keep
 #include <emscripten.h>
-#include <emscripten/em_js.h>
 // IWYU pragma: end_keep
 
 #include "platforms/coroutine_runtime.h"
+#include "fl/stl/condition_variable.h"
 #include "fl/stl/int.h"
+#include "fl/stl/mutex.h"
 #include "fl/stl/noexcept.h"
+#include "fl/stl/thread.h"
+#include "fl/stl/unique_ptr.h"
+
+// Forward-declare pthread_exit (stable POSIX C ABI). Avoids pulling
+// <pthread.h> into a header — that include is banned project-wide in favor
+// of fl/stl/thread.h. We only need this one symbol to terminate the worker
+// thread cleanly during destroyContext() teardown; see contextSwitch().
+extern "C" __attribute__((noreturn)) void pthread_exit(void* retval);
 
 namespace fl {
 namespace platforms {
 
-static constexpr int kMaxJspiContexts = 32;
+/// Per-context state for the pthread-based coroutine platform.
+///
+/// Lifetime: the platform owns this struct directly. The worker thread holds
+/// only a raw `self` pointer, but contextSwitch() observes `exit_requested`
+/// after every wake and calls `pthread_exit()` so the worker is fully torn
+/// down (with no further access to *this) before destroyContext() proceeds
+/// with `join()`. See destroyContext() comments for the teardown protocol.
+struct PthreadCoroCtx {
+    fl::unique_ptr<fl::thread> thread;     ///< null for the runner context (no OS thread)
+    fl::mutex mutex;                        ///< Guards `signaled` / `exit_requested`
+    fl::condition_variable cv;              ///< Park spot for this thread
+    bool signaled = false;                  ///< Set true to wake this context
+    bool exit_requested = false;            ///< Set true to terminate the thread
+    void (*entry_fn)() = nullptr;           ///< User-supplied entry trampoline
+};
 
-/// Entry function table — indexed by context ID.
-/// ID 0 is reserved for the runner (main thread) context.
-static void (*g_jspi_entries[kMaxJspiContexts])() = {};
-static int g_jspi_next_id = 1;
-
-}  // namespace platforms
-}  // namespace fl
-
-// ==========================================================================
-// WASM export: coroutine entry point (called from JS scheduler)
-// ==========================================================================
-// Must be in JSPI_EXPORTS so the engine wraps it as a promising export.
-// When called, the WASM runs synchronously until the coroutine's first yield.
-
-extern "C" EMSCRIPTEN_KEEPALIVE void jspi_run_entry(int id) {
-    using namespace fl::platforms;
-    if (id >= 0 && id < kMaxJspiContexts && g_jspi_entries[id]) {
-        g_jspi_entries[id]();
-    }
-}
-
-// ==========================================================================
-// JS scheduler glue — forward declarations (defined via EM_JS in
-// coroutine_platform_wasm.js.cpp.hpp, included from _build.cpp.hpp)
-// ==========================================================================
-extern "C" {
-void _jspi_register(int id);
-void _jspi_unregister(int id);
-void _jspi_context_switch(int from_id, int to_id);
-}
-
-// ==========================================================================
-// CoroutinePlatformWasm — ICoroutinePlatform implementation
-// ==========================================================================
-
-namespace fl {
-namespace platforms {
-
-// NOLINTBEGIN(reinterpret_cast)
-// Context IDs are small ints stored in void* — reinterpret_cast is unavoidable
-// because the ICoroutinePlatform interface uses void* for platform contexts.
-
-inline void* id_to_ctx(int id) FL_NOEXCEPT {
-    return reinterpret_cast<void*>(static_cast<uintptr_t>(id));  // ok reinterpret cast
-}
-
-inline int ctx_to_id(void* ctx) FL_NOEXCEPT {
-    return static_cast<int>(reinterpret_cast<uintptr_t>(ctx));  // ok reinterpret cast
-}
-// NOLINTEND(reinterpret_cast)
-
-class CoroutinePlatformWasm : public ICoroutinePlatform {
-public:
-    void* createContext(void (*entry_fn)(), size_t /*stack_size*/) override {
-        // JSPI manages stacks at the engine level — stack_size is ignored
-        int id = g_jspi_next_id++;
-        if (id >= kMaxJspiContexts) {
-            return nullptr;
+/// Thread body. Each non-runner context owns one of these.
+/// Park until first signal, then run the entry trampoline. The trampoline
+/// itself never returns (CoroutineContext::entry_trampoline busy-loops if
+/// it falls through), so the *only* way this function returns is via the
+/// initial-wait exit branch below. Mid-coroutine teardown is handled by
+/// contextSwitch() calling pthread_exit() — the thread terminates without
+/// returning here.
+inline void pthread_coro_thread_main(PthreadCoroCtx* self) FL_NOEXCEPT {
+    {
+        fl::unique_lock<fl::mutex> lk(self->mutex);
+        self->cv.wait(lk, [self] { return self->signaled; });
+        self->signaled = false;
+        if (self->exit_requested) {
+            // Initial-wait exit — destroyContext() called before the
+            // coroutine ever ran. Drop the lock and exit cleanly so
+            // destroyContext()'s join() can return.
+            return;
         }
-        g_jspi_entries[id] = entry_fn;
-        _jspi_register(id);
-        return id_to_ctx(id);
+    }
+
+    if (self->entry_fn) {
+        self->entry_fn();
+    }
+    // Unreachable: entry_fn (CoroutineContext::entry_trampoline) busy-loops
+    // on its own after the final contextSwitch back to the runner.
+    // Mid-coroutine teardown exits the thread via pthread_exit() inside
+    // contextSwitch(), so control never returns here from entry_fn().
+}
+
+class CoroutinePlatformPthread : public ICoroutinePlatform {
+public:
+    void* createContext(void (*entry_fn)(), size_t /*stack_size*/) FL_NOEXCEPT override {
+        auto* ctx = new PthreadCoroCtx();  // ok bare allocation - platform context lifetime
+        ctx->entry_fn = entry_fn;
+        ctx->thread.reset(new fl::thread(pthread_coro_thread_main, ctx));  // ok bare allocation
+        return ctx;
     }
 
     void* createRunnerContext() FL_NOEXCEPT override {
-        // Runner is always ID 0
-        _jspi_register(0);
-        return id_to_ctx(0);
+        // Runner is the calling thread (sketch thread). No OS thread to spawn —
+        // we just need a parking spot for it.
+        auto* ctx = new PthreadCoroCtx();  // ok bare allocation
+        return ctx;
     }
 
-    void destroyContext(void* ctx) FL_NOEXCEPT override {
-        int id = ctx_to_id(ctx);
-        if (id >= 0 && id < kMaxJspiContexts) {
-            g_jspi_entries[id] = nullptr;
-            _jspi_unregister(id);
+    void destroyContext(void* ctx_p) FL_NOEXCEPT override {
+        auto* ctx = static_cast<PthreadCoroCtx*>(ctx_p);
+        if (!ctx) return;
+
+        if (ctx->thread) {
+            // Teardown protocol — guarantees the worker has fully stopped
+            // touching *ctx before we free it.
+            //
+            //   1. Set `exit_requested` and `signaled` under the lock.
+            //   2. notify_all() — wakes the worker out of either its initial
+            //      cv.wait() or any contextSwitch() cv.wait().
+            //   3. The worker rechecks `exit_requested` after every cv.wait()
+            //      and either falls off pthread_coro_thread_main (initial
+            //      wait) or calls pthread_exit() (contextSwitch).
+            //   4. join() blocks until that exit completes — only then is it
+            //      safe to delete the thread object and *ctx.
+            {
+                fl::lock_guard<fl::mutex> lk(ctx->mutex);
+                ctx->exit_requested = true;
+                ctx->signaled = true;
+            }
+            ctx->cv.notify_all();
+
+            if (ctx->thread->joinable()) {
+                ctx->thread->join();
+            }
+            ctx->thread.reset();
         }
+        delete ctx;  // ok bare allocation - platform context lifetime
     }
 
-    void contextSwitch(void* from_ctx, void* to_ctx) FL_NOEXCEPT override {
-        _jspi_context_switch(ctx_to_id(from_ctx), ctx_to_id(to_ctx));
+    void contextSwitch(void* from_ctx_p, void* to_ctx_p) FL_NOEXCEPT override {
+        auto* from = static_cast<PthreadCoroCtx*>(from_ctx_p);
+        auto* to = static_cast<PthreadCoroCtx*>(to_ctx_p);
+        if (!from || !to) return;
+
+        // Pass the baton to `to`.
+        {
+            fl::lock_guard<fl::mutex> lk(to->mutex);
+            to->signaled = true;
+        }
+        to->cv.notify_one();
+
+        // Park until someone hands the baton back to us — or until
+        // destroyContext() asks us to exit (`exit_requested`).
+        bool exit_now = false;
+        {
+            fl::unique_lock<fl::mutex> lk(from->mutex);
+            from->cv.wait(lk, [from] { return from->signaled; });
+            from->signaled = false;
+            exit_now = from->exit_requested;
+        }
+        if (exit_now) {
+            // Mid-coroutine teardown. The trampoline's caller frame would
+            // otherwise busy-loop (entry_trampoline's `while (true) {}`),
+            // pinning the worker forever and blocking destroyContext()'s
+            // join(). Terminate the thread cleanly here — no further access
+            // to *from after this point, so destroyContext() may free *from
+            // as soon as join() returns.
+            pthread_exit(nullptr);
+        }
     }
 
     fl::u32 micros() const FL_NOEXCEPT override {
@@ -126,4 +177,4 @@ public:
 }  // namespace platforms
 }  // namespace fl
 
-#endif  // FL_IS_WASM && !FASTLED_WASM_PTHREADS
+#endif  // FL_IS_WASM
